+++ cognite/client/data_classes/sequences.py
@@ -479,7 +479,7 @@ class SequenceColumnUpdate(CogniteUpdate):
         return SequenceColumnUpdate._ObjectSequenceColumnUpdate(self, "metadata")
 
     @classmethod
-    def _get_update_properties(cls) -> list[PropertySpec]:
+    def _get_update_properties(cls, item: CogniteResource | None = None) -> list[PropertySpec]:
         return [
             PropertySpec("description"),
             PropertySpec("external_id", is_nullable=False),
@@ -578,7 +578,7 @@ class SequenceUpdate(CogniteUpdate):
         return SequenceUpdate._ColumnsSequenceUpdate(self, "columns")
 
     @classmethod
-    def _get_update_properties(cls) -> list[PropertySpec]:
+    def _get_update_properties(cls, item: CogniteResource | None = None) -> list[PropertySpec]:
         return [
             # External ID is nullable, but is used in the upsert logic and thus cannot be nulled out.
             PropertySpec("external_id", is_nullable=False),