+++ CHANGELOG.md
@@ -17,6 +17,14 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [7.53.4] - 2024-07-11
+### Added
+- `FilesAPI.upload_bytes` and `FilesAPI.upload` are updated to be compatible with Private Link projects.
+
+## [7.53.3] - 2024-07-11
+### Added
+- [Feature Preview - alpha] Support for `instanceId` in the `client.time_series` `.retrieve`, `.retrieve_multiple`,
+  and `.update` methods. This is an experimental feature and may change without warning.
 
 ## [7.53.2] - 2024-07-03
 ### Fixed